Implement XbitTextAppGateway
Scan for text message advertisements from devices.
Only allow a message from a device every 5 seconds.
Only allow new messages from a device. If the same message
is receive from a device, it is ignored.

Flash blue LED when a new message is received from a device.
Flash the green LED on a successful publish.

PROD-7174

Signed-off-by: Ryan Erickson <ryan.erickson@lairdconnect.com>