AT command updates for Click board
This change allows the AT command script to run on the NX040 click board. The Click board has the Zephyr shell and MicroPython REPL on the USB bus of the Click board, but the AT command interface needs to use UART0.

The main script was updated to detect the board type and use stdio for the DVK and UART0 for the Click board. All input and output go through a new interface to the ATReader class.

The ATReader class was updated to handle both string and bytes input from the read() function. The stdio interface returns strings, but the UART interface returns bytes. There's also handling for receiving \r instead of \n as the line terminator.

PROD-10158