Initial commit of restructured Canvas Python Samples
This (large) commit replaces the prior sample folders/files with a  more organized and concise list of Python sample script code organized into 3 categories (demos, snippets and starter projects). This is part of the ongoing effort to improve the quality of the Canvas developer experience and is the first step toward improving organization and maintainability of the samples repository.

PROD-16404