Added button_led demo for Veda SL917
This demo is linked to by the "Veda SL917 Development with Canvas" presentation and video being developed as part of the 2025 workshop content in collaboration with SiLabs. It provides a simple example to beginners of how to write a Python application on the Veda SL917 SoC explorer board. Also removing the .gitignore per Jonathan's request as it is confusing to users not familiar with the file to see it on the "home page" of the Canvas Python Samples repository on our GitHub site (please use a local copy of .gitignore as desired for your environment).

PROD-23609